backend/requirements.txt
Flask
Flask-SQLAlchemy
pytest
python-dotenv
requests
Flask-Caching
gunicorn 